python/requirements.txt
Flask==1.0.2
Flask-Cors==3.0.4
mecab-python3==0.7